requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
pandas
requests
python-dotenv
psycopg2-binary
